add option 'keep_maxop_trajs' in [output]

+++ infretis/classes/repex.py
@@ -163,6 +163,16 @@ class REPEX_state:
         """Retrieve workers from config dict."""
         return self.config["runner"]["workers"]
 
+    @property
+    def maxop(self):
+        """Get the maximum orderparameter seen during the simulation."""
+        return self.config["current"]["maxop"]
+
+    @maxop.setter
+    def maxop(self, val):
+        """Update the maximum orderpameter seen druing the sumulation."""
+        self.config["current"]["maxop"] = val
+
     def pick(self):
         """Pick path and ens."""
         prob = self.prob.astype("float64").flatten()
@@ -895,6 +905,9 @@ class REPEX_state:
         for ens_num in picked.keys():
             pn_old = picked[ens_num]["pn_old"]
             out_traj = picked[ens_num]["traj"]
+            # keep track of the highest order value seen during the sim
+            if ens_num != -1 and out_traj.ordermax[0] > self.maxop:
+                self.maxop = out_traj.ordermax[0]
             self.ensembles[ens_num + 1] = picked[ens_num]["ens"]
 
             for idx, lock in enumerate(self.locked):
@@ -928,9 +941,16 @@ class REPEX_state:
                 ):
                     if len(self.pn_olds) > self.n - 2:
                         pn_old_del, del_dic = next(iter(self.pn_olds.items()))
-                        # delete trajectory files
-                        for adress in del_dic["adress"]:
-                            os.remove(adress)
+                        if self.config["output"]["keep_maxop_trajs"]:
+                            # delete trajectory files if low orderp (infinit)
+                            if del_dic["max_op"][0] < self.maxop:
+                                # update maxop and then delete|
+                                for adress in del_dic["adress"]:
+                                    os.remove(adress)
+                        else:
+                            # delete trajectory files
+                            for adress in del_dic["adress"]:
+                                os.remove(adress)
                         # delete txt files
                         load_dir = self.config["simulation"]["load_dir"]
                         if self.config["output"].get("delete_old_all", False):
@@ -950,6 +970,7 @@ class REPEX_state:
                     if len(self.pn_olds) <= self.n - 2:
                         self.pn_olds[str(pn_old)] = {
                             "adress": self.traj_data[pn_old]["adress"],
+                            "max_op": self.traj_data[pn_old]["max_op"],
                         }
             pn_news.append(out_traj.path_number)
             self.add_traj(ens_num, out_traj, valid=out_traj.weights)
@@ -1009,6 +1030,9 @@ class REPEX_state:
                 "weights": paths[i + 1].weights,
                 "frac": np.array(frac, dtype="longdouble"),
             }
+            # keep track of max op of i+ path
+            if paths[i + 1].ordermax[0] > self.maxop:
+                self.maxop = paths[i + 1].ordermax[0]
         # add minus path:
         paths[0].weights = (1.0,)
         pnum = paths[0].path_number
